requirements.txt
-i https://pypi.org/simple
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
boto3==1.24.61
botocore==1.27.61
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.0.4
cycler==0.11.0
decorator==5.1.1
executing==1.0.0
fastapi==0.81.0
fire==0.4.0
flatbuffers==1.12
fonttools==4.37.1
frozenlist==1.3.1
fsspec==2022.7.1
gast==0.4.0
gcsfs==2022.7.1
google-api-core==2.8.2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
h11==0.13.0
h5py==3.7.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
jmespath==1.0.1
joblib==1.1.0
keras==2.9.0
keras-preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
markdown==3.4.1
markupsafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
multidict==6.0.2
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.4
pydantic==1.9.2
pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pyyaml==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
sahi==0.10.4
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.11.2
setuptools==65.3.0
shapely==1.8.4
six==1.16.0
sniffio==1.2.0
stack-data==0.5.0
starlette==0.19.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminaltables==3.1.10
thop==0.1.1.post2207130030
threadpoolctl==3.1.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
traitlets==5.3.0
typing-extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
wcwidth==0.2.5
werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
yarl==1.8.1
yolov5==6.1.8
